Rename the test data command to the template filename
The cookiecutter template scaffolds `generate_nautobot_golden_config_test_data`,
which collided with this app's pre-existing `generate_gc_test_data` and left two
`generate_test_data` invoke tasks. Renaming the real command to the template's
filename means future drift updates land on the same file instead of re-adding a
stub and a duplicate task.

- git mv generate_gc_test_data.py -> generate_nautobot_golden_config_test_data.py
  (replacing the template's stub)
- tasks.py: `invoke generate-test-data` now calls the renamed command
- docs/dev/dev_environment.md: updated the documented command name
